@@ -1,60 +1,45 @@
 {
-	"name": "frontend",
+	"name": "svelte5-frontend",
 	"version": "0.0.1",
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"test": "vitest",
-		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write .",
-		"studio": "drizzle-kit studio",
-		"generate": "drizzle-kit generate",
-		"migrate": "drizzle-kit migrate"
+		"lint": "prettier --check . && eslint ."
 	},
 	"devDependencies": {
 		"@fontsource/fira-mono": "^5.0.0",
 		"@neoconfetti/svelte": "^2.0.0",
-		"@sveltejs/adapter-auto": "^3.2.3",
-		"@sveltejs/kit": "^2.5.21",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
+		"@prisma/client": "^5.22.0",
+		"@sveltejs/adapter-auto": "^3.0.0",
+		"@sveltejs/kit": "^2.0.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/eslint": "^9.6.0",
 		"autoprefixer": "^10.4.20",
-		"dotenv": "^16.4.5",
-		"drizzle-kit": "^0.23.1-7816536",
-		"eslint": "^9.0.0",
+		"bits-ui": "^1.0.0-next.63",
+		"clsx": "^2.1.1",
+		"eslint": "^9.7.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.36.0",
 		"globals": "^15.0.0",
-		"pg": "^8.12.0",
-		"postcss": "^8.4.40",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"svelte": "^4.2.18",
-		"svelte-check": "^3.6.0",
-		"svelte-headless-table": "^0.18.2",
-		"tailwindcss": "^3.4.7",
+		"lucide-svelte": "^0.460.1",
+		"postcss": "^8.4.49",
+		"prettier": "^3.3.2",
+		"prettier-plugin-svelte": "^3.2.6",
+		"prettier-plugin-tailwindcss": "^0.6.5",
+		"prisma": "^5.22.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"tailwind-merge": "^2.5.4",
+		"tailwind-variants": "^0.3.0",
+		"tailwindcss": "^3.4.15",
+		"tailwindcss-animate": "^1.0.7",
 		"typescript": "^5.0.0",
 		"typescript-eslint": "^8.0.0",
-		"vite": "^5.0.3",
-		"vitest": "^2.0.0"
-	},
-	"type": "module",
-	"dependencies": {
-		"axios": "^1.7.3",
-		"bits-ui": "^0.21.13",
-		"clsx": "^2.1.1",
-		"drizzle-orm": "^0.32.1",
-		"formsnap": "^1.0.1",
-		"lucide-svelte": "^0.424.0",
-		"mode-watcher": "^0.4.1",
-		"postgres": "^3.4.4",
-		"svelte-sonner": "^0.3.27",
-		"sveltekit-superforms": "^2.16.1",
-		"tailwind-merge": "^2.4.0",
-		"tailwind-variants": "^0.2.1",
-		"zod": "^3.23.8"
+		"vite": "^5.0.3"
 	}
 }
